refactor: make instructions module and move ixs to it
also some renames and clean up imports
